feature: created useNearbyFires hook
